+++ src/Core/VRChatContentPublisher.Core/Services/PublishTask/ContentPublisher/WorldContentPublisher.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.Extensions.Logging;
+using VRChatContentPublisher.ConnectCore.Models;
 using VRChatContentPublisher.ConnectCore.Services;
 using VRChatContentPublisher.Core.Models;
 using VRChatContentPublisher.Core.Models.VRChatApi;
@@ -10,58 +11,81 @@ using VRChatContentPublisher.Core.Utils;
 
 namespace VRChatContentPublisher.Core.Services.PublishTask.ContentPublisher;
 
-public sealed class WorldContentPublisher : IContentPublisher
+public sealed class WorldContentPublisher(
+    string worldId,
+    string worldName,
+    string platform,
+    string unityVersion,
+    string? worldSignature,
+    int? capacity,
+    int? recommendedCapacity,
+    string? previewYoutubeId,
+    string[]? udonProducts,
+    UserSessionService userSessionService,
+    ILogger<WorldContentPublisher> logger,
+    IFileService tempFileService)
+    : IContentPublisher
 {
-    private readonly string _worldId;
-    private readonly string _worldName;
-    private readonly string _platform;
-    private readonly string _unityVersion;
-    private readonly string? _worldSignature;
-    private readonly int? _capacity;
-    private readonly int? _recommendedCapacity;
-    private readonly string? _previewYoutubeId;
-    private readonly string[] _udonProducts;
+    private readonly string[] _udonProducts = udonProducts ?? [];
 
-    private readonly UserSessionService _userSessionService;
-
-    private readonly ILogger<WorldContentPublisher> _logger;
-    private readonly IFileService _tempFileService;
-
-    private readonly VRChatApiClient _apiClient;
-
-    public WorldContentPublisher(string worldId,
-        string worldName,
-        string platform,
-        string unityVersion,
-        string? worldSignature,
-        int? capacity,
-        int? recommendedCapacity,
-        string? previewYoutubeId,
-        string[]? udonProducts,
-        UserSessionService userSessionService,
-        ILogger<WorldContentPublisher> logger,
-        IFileService tempFileService)
-    {
-        _worldId = worldId;
-        _worldName = worldName;
-        _platform = platform;
-        _unityVersion = unityVersion;
-        _worldSignature = worldSignature;
-        _userSessionService = userSessionService;
-        _logger = logger;
-        _tempFileService = tempFileService;
-        _capacity = capacity;
-        _recommendedCapacity = recommendedCapacity;
-        _previewYoutubeId = previewYoutubeId;
-        _udonProducts = udonProducts ?? [];
-
-        _apiClient = _userSessionService.GetApiClient();
-    }
+    private readonly VRChatApiClient _apiClient = userSessionService.GetApiClient();
 
     public string GetContentType() => "world";
 
-    public string GetContentName() => _worldName;
-    public string GetContentPlatform() => _platform;
+    public string GetContentName() => worldName;
+    public string GetContentPlatform() => platform;
+
+    public async ValueTask BeforePublishTaskAsync(
+        string? thumbnailFileId,
+        string? description,
+        string[]? tags,
+        string? releaseStatus,
+        HttpClient awsClient,
+        CancellationToken cancellationToken = default
+    )
+    {
+        // try fetch world detail, if not found means we need to create a new world.
+        try
+        {
+            await _apiClient.GetWorldAsync(worldId);
+            return;
+        }
+        catch (ApiErrorException ex) when (ex.StatusCode == 404)
+        {
+            logger.LogInformation("The world {WorldId} was not found. Creating new world.", worldId);
+        }
+
+        logger.LogInformation("Uploading thumbnail file for creating new world {WorldId}", worldId);
+        if (thumbnailFileId is null)
+            throw new InvalidOperationException("Thumbnail must be provided when creating a new world.");
+
+        var thumbnailFile = await tempFileService.GetFileWithNameAsync(thumbnailFileId);
+        await using var thumbnailFileStream = thumbnailFile?.FileStream;
+
+        if (thumbnailFile is null || thumbnailFileStream is null)
+            throw new ArgumentException("Could not find the provided thumbnail file.", nameof(thumbnailFileId));
+
+        var imageUrl = await UploadThumbnailFileAsync(null, thumbnailFile, awsClient, null, cancellationToken);
+        
+        logger.LogInformation("Send create world request for {WorldId}", worldId);
+        await _apiClient.CreateWorldAsync(new CreateWorldRequest(
+            worldId,
+            worldName,
+            null,
+            null,
+            null,
+            null,
+            null,
+            imageUrl,
+            description,
+            tags,
+            releaseStatus,
+            capacity,
+            recommendedCapacity,
+            previewYoutubeId,
+            null
+        ), cancellationToken);
+    }
 
     private const long MaxBundleFileSizeForMobileBytes = 104857600; // 100 MB
 
@@ -75,9 +99,9 @@ public sealed class WorldContentPublisher : IContentPublisher
         PublishStageProgressReporter? progressReporter = null,
         CancellationToken cancellationToken = default)
     {
-        await using var bundleFileStream = await _tempFileService.GetFileAsync(bundleFileId);
+        await using var bundleFileStream = await tempFileService.GetFileAsync(bundleFileId);
         var thumbnailFile = thumbnailFileId is not null
-            ? await _tempFileService.GetFileWithNameAsync(thumbnailFileId)
+            ? await tempFileService.GetFileWithNameAsync(thumbnailFileId)
             : null;
         await using var thumbnailFileStream = thumbnailFile?.FileStream;
 
@@ -87,7 +111,7 @@ public sealed class WorldContentPublisher : IContentPublisher
         if (thumbnailFile is null && thumbnailFileId is not null)
             throw new ArgumentException("Could not find the provided thumbnail file.", nameof(thumbnailFileId));
 
-        if (!UnityBuildTargetUtils.IsStandalonePlatform(_platform) &&
+        if (!UnityBuildTargetUtils.IsStandalonePlatform(platform) &&
             bundleFileStream.Length > MaxBundleFileSizeForMobileBytes)
             throw new ArgumentException(
                 "The provided bundle file exceeds the maximum allowed size of 100 MB for this platform.",
@@ -95,25 +119,17 @@ public sealed class WorldContentPublisher : IContentPublisher
 
         cancellationToken.ThrowIfCancellationRequested();
 
-        _logger.LogInformation("Publish World {WorldId}", _worldId);
+        logger.LogInformation("Publish World {WorldId}", worldId);
         progressReporter?.Report("Fetching world detail...");
 
-        // Step 1. Fetch world detail, if not found means we need to create a new world.
-        VRChatApiWorld? world = null;
-        try
-        {
-            world = await _apiClient.GetWorldAsync(_worldId);
-        }
-        catch (ApiErrorException ex) when (ex.StatusCode == 404)
-        {
-            _logger.LogInformation("The world {WorldId} was not found. Will create new world.", _worldId);
-        }
+        // Step 1. Fetch world detail, it should be always exist since pre publish will ensure this.
+        var world = await _apiClient.GetWorldAsync(worldId);
 
         // Step 2. Try to get the asset file for this platform, if not create a new one.
         // This step also cleanups any incomplete file versions.
         var fileId = await GetOrCreateBundleFileIdAsync(world);
 
-        _logger.LogInformation("Using file id {FileId} for world {WorldId}", fileId, _worldId);
+        logger.LogInformation("Using file id {FileId} for world {WorldId}", fileId, worldId);
         progressReporter?.Report("Preparing for upload bundle file...");
 
         // Step 3. Create and upload a new file version
@@ -131,86 +147,45 @@ public sealed class WorldContentPublisher : IContentPublisher
         string? imageUri = null;
         if (thumbnailFile is not null && thumbnailFileStream is not null)
         {
-            _logger.LogInformation("Uploading thumbnail for avatar {AvatarId}", _worldId);
+            logger.LogInformation("Uploading thumbnail for world {AvatarId}", worldId);
             progressReporter?.Report("Uploading thumbnail...");
 
-            var imageFileId = await GetOrCreateBundleImageIdAsync(world, thumbnailFile.FileName);
-
-            var imageFileVersion = await _apiClient.CreateAndUploadFileVersionAsync(
-                thumbnailFileStream,
-                imageFileId,
-                VRChatApiFlieUtils.GetMimeTypeFromExtension(Path.GetExtension(thumbnailFile.FileName)),
-                awsClient,
-                "World Thumbnail",
-                arg => progressReporter?.Report(arg.ProgressText, arg.ProgressValue), cancellationToken
-            );
-
-            if (imageFileVersion.File is null)
-                throw new UnexpectedApiBehaviourException(
-                    "Api did not return file info for created image file version.");
-
-            imageUri = imageFileVersion.File.Url;
+            imageUri = await UploadThumbnailFileAsync(world, thumbnailFile, awsClient, progressReporter,
+                cancellationToken);
         }
 
         if (fileVersion.File is null)
             throw new UnexpectedApiBehaviourException("Api did not return file info for created file version.");
 
-        // Step 4. Update or Create World
-        if (world is not null)
-        {
-            _logger.LogInformation("Updating world {WorldId} to use new file version {Version}", _worldId,
-                fileVersion.Version);
-            progressReporter?.Report("Updating world to latest asset version...");
+        // Step 4. Update World
+        logger.LogInformation("Updating world {WorldId} to use new file version {Version}", worldId,
+            fileVersion.Version);
+        progressReporter?.Report("Updating world to latest asset version...");
 
-            await _apiClient.CreateWorldVersionAsync(_worldId, new CreateWorldVersionRequest(
-                _worldName,
-                fileVersion.File.Url,
-                fileVersion.Version,
-                _platform,
-                _unityVersion,
-                _worldSignature,
-                imageUri,
-                description,
-                tags,
-                releaseStatus,
-                _capacity,
-                _recommendedCapacity,
-                _previewYoutubeId,
-                _udonProducts
-            ), cancellationToken);
-        }
-        else
-        {
-            _logger.LogInformation("Creating new world {WorldId} with file version {Version}", _worldId,
-                fileVersion.Version);
-            progressReporter?.Report("Creating new world...");
+        await _apiClient.CreateWorldVersionAsync(worldId, new CreateWorldVersionRequest(
+            worldName,
+            fileVersion.File.Url,
+            fileVersion.Version,
+            platform,
+            unityVersion,
+            worldSignature,
+            imageUri,
+            description,
+            tags,
+            releaseStatus,
+            capacity,
+            recommendedCapacity,
+            previewYoutubeId,
+            _udonProducts
+        ), cancellationToken);
 
-            await _apiClient.CreateWorldAsync(new CreateWorldRequest(
-                _worldId,
-                _worldName,
-                fileVersion.File.Url,
-                4,
-                _platform,
-                _unityVersion,
-                _worldSignature,
-                imageUri,
-                description,
-                tags,
-                releaseStatus,
-                _capacity,
-                _recommendedCapacity,
-                _previewYoutubeId,
-                _udonProducts
-            ), cancellationToken);
-        }
-
-        _logger.LogInformation("Successfully published world {WorldId}", _worldId);
+        logger.LogInformation("Successfully published world {WorldId}", worldId);
     }
 
     private VRChatApiUnityPackage? TryGetUnityPackageForPlatform(VRChatApiWorld world)
     {
         return world.UnityPackages
-            .Where(package => package.Platform == _platform)
+            .Where(package => package.Platform == platform)
             .GroupBy(package => package.UnityVersion)
             .MaxBy(group => UnityVersion.TryParse(group.Key))?
             .MaxBy(package => package.AssetVersion);
@@ -231,9 +206,35 @@ public sealed class WorldContentPublisher : IContentPublisher
             }
         }
 
-        var fileName = $"World - {_worldName} - Asset bundle - {_unityVersion}-{_platform}";
+        var fileName = $"World - {worldName} - Asset bundle - {unityVersion}-{platform}";
         var file = await _apiClient.CreateFileAsync(fileName, "application/x-world", ".vrcw");
         return file.Id;
+    }
+
+    private async ValueTask<string> UploadThumbnailFileAsync(
+        VRChatApiWorld? world,
+        UploadedFile thumbnailFile,
+        HttpClient awsClient,
+        PublishStageProgressReporter? progressReporter,
+        CancellationToken cancellationToken = default
+    )
+    {
+        var imageFileId = await GetOrCreateBundleImageIdAsync(world, thumbnailFile.FileName);
+
+        var imageFileVersion = await _apiClient.CreateAndUploadFileVersionAsync(
+            thumbnailFile.FileStream,
+            imageFileId,
+            VRChatApiFlieUtils.GetMimeTypeFromExtension(Path.GetExtension(thumbnailFile.FileName)),
+            awsClient,
+            "World Thumbnail",
+            arg => progressReporter?.Report(arg.ProgressText, arg.ProgressValue), cancellationToken
+        );
+
+        if (imageFileVersion.File is null)
+            throw new UnexpectedApiBehaviourException(
+                "Api did not return file info for created image file version.");
+
+        return imageFileVersion.File.Url;
     }
 
     private async ValueTask<string> GetOrCreateBundleImageIdAsync(VRChatApiWorld? apiWorld, string imageFileName)
@@ -250,7 +251,7 @@ public sealed class WorldContentPublisher : IContentPublisher
         var extension = Path.GetExtension(imageFileName);
         var mimeType = VRChatApiFlieUtils.GetMimeTypeFromExtension(extension);
 
-        var fileName = $"World - {_worldName} - Image - {_unityVersion}-{_platform}";
+        var fileName = $"World - {worldName} - Image - {unityVersion}-{platform}";
         var file = await _apiClient.CreateFileAsync(fileName, mimeType, extension);
         return file.Id;
     }